[MS] Added Responsive Menu
