+++ client/src/theme/components/modals.scss
@@ -1,5 +1,15 @@
 /* Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS */
 
+@import '@/theme/responsive-mixin';
+
+/* -- modal-global -- */
+@include breakpoint('sm') {
+  ion-modal::part(content) {
+    width: 100%;
+    min-width: 100%;
+  }
+}
+
 /* -- modal-stepper -- */
 .modal-stepper {
   // global
