+++ client/src/theme/global.scss
@@ -10,6 +10,7 @@
 @import 'components/notifications';
 @import 'components/popovers';
 @import 'components/tags';
+@import '@/theme/responsive-mixin';
 @import 'global-client';
 
 /**** Global CSS overloads ****/
