+++ client/src/theme/responsive-mixin.scss
@@ -1,0 +1,37 @@
+/* Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS */
+
+$breakpoints: (
+  'xs': (
+    max-width: 576px,
+  ),
+  'sm': (
+    max-width: 768px,
+  ),
+  'md': (
+    max-width: 992px,
+  ),
+  'lg': (
+    max-width: 1200px,
+  ),
+  'wide': (
+    max-width: 1500px,
+  ),
+  'ultra-wide': (
+    max-width: 1921px,
+  ),
+) !default;
+
+@mixin breakpoint($breakpoint) {
+  // If the key exists in the map
+  @if map-has-key($breakpoints, $breakpoint) {
+    // Prints a media query based on the value
+    @media #{inspect(map-get($breakpoints, $breakpoint))} {
+      @content;
+    }
+  }
+  // If the key doesn't exist in the map
+  @else {
+    @warn "Unfortunately, no value could be retrieved from `#{$breakpoint}`. "
+        + "Available breakpoints are: #{map-keys($breakpoints)}.";
+  }
+}
