+++ client/src/views/menu/index.ts
@@ -1,5 +1,5 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
 
-import { SidebarSavedData } from '@/views/sidebar-menu/utils';
+import { SidebarSavedData } from '@/views/menu/utils';
 
 export { SidebarSavedData };
