+++ newsfragments/9805.feature.rst
@@ -1,0 +1,1 @@
+Switch sidebar to tab menu on mobile
